gateway/requirements.txt
fastapi
slowapi
uvicorn
httpx
pyjwt
python-dotenv
python-multipart